chore: everything except styles
