+++ packages/components/forms/src/select/Select.mdx
@@ -5,6 +5,7 @@ status: 'stable'
 slug: /components/select/
 github: 'https://github.com/contentful/forma-36/tree/master/packages/forma-36-react-components/src/components/Select'
 storybook: 'https://f36-storybook.contentful.com/?path=/story/form-elements-select--overview'
+typescript: ./Select.tsx,./SelectOption.tsx
 ---
 
 Select allows users to make a single selection from an options menu.
@@ -34,48 +35,48 @@ Keep in mind that Select should not be used without the label. We provide [Selec
 ## Code examples
 
 ```tsx
-// import { Select } from '@contentful/forma-36-react-components';
+// import { Select } from '@contentful/f36-components';
 
 <Select id="optionSelect" name="optionSelect" width="large">
-  <Option value="optionOne">Option 1</Option>
-  <Option value="optionTwo">Long Option 2</Option>
+  <Select.Option value="optionOne">Option 1</Select.Option>
+  <Select.Option value="optionTwo">Long Option 2</Select.Option>
 </Select>
 ```
 
 ### Select state (`disabled`)
 
 ```tsx
-// import { Select } from '@contentful/forma-36-react-components';
+// import { Select } from '@contentful/f36-components';
 
 <Select id="optionSelect" name="optionSelect" isDisabled>
-  <Option value="optionOne">Option 1</Option>
-  <Option value="optionTwo">Long Option 2</Option>
+  <Select.Option value="optionOne">Option 1</Select.Option>
+  <Select.Option value="optionTwo">Long Option 2</Select.Option>
 </Select>
 ```
 
 ### Select with placeholder
 
 ```tsx
-// import { Select } from '@contentful/forma-36-react-components';
+// import { Select } from '@contentful/f36-components';
 
 <Select name="optionSelect" id="optionSelect">
-  <Option value="" disabled selected>
+  <Select.Option value="" disabled selected>
     Please select an option...
-  </Option>
-  <Option value="optionOne">Option One</Option>
-  <Option value="optionTwo">Option Two</Option>
-  <Option value="optionThree">Option Three</Option>
+  </Select.Option>
+  <Select.Option value="optionOne">Option One</Select.Option>
+  <Select.Option value="optionTwo">Option Two</Select.Option>
+  <Select.Option value="optionThree">Option Three</Select.Option>
 </Select>
 ```
 
 ### Select with error
 
 ```tsx
-// import { Select } from '@contentful/forma-36-react-components';
+// import { Select } from '@contentful/f36-components';
 
 <Select id="optionSelect" name="optionSelect" hasError>
-  <Option value="optionOne">Option 1</Option>
-  <Option value="optionTwo">Long Option 2</Option>
+  <Select.Option value="optionOne">Option 1</Select.Option>
+  <Select.Option value="optionTwo">Long Option 2</Select.Option>
 </Select>
 ```
 
@@ -88,3 +89,15 @@ Keep in mind that Select should not be used without the label. We provide [Selec
 
 - Select should contain a label, if needed help message and validation message if needed. We provide [SelectField](https://f36.contentful.com/components/select-field/) component, that contains all of those elements.
 - Select should always a `name` property.
+
+## Props
+
+import { Props } from '@contentful/f36-docs-utils';
+
+### Select
+
+<Props of="Select" />
+
+### Select.Option
+
+<Props of="Option" />
