+++ packages/components/forms/src/select/index.ts
@@ -1,4 +1,4 @@
 export { Option } from './SelectOption';
 export type { OptionProps } from './SelectOption';
-export { Select } from './Select';
+export { Select } from './CompoundSelect';
 export type { SelectInternalProps, SelectProps } from './Select';
