+++ packages/components/forms/src/text-input/TextInput.mdx
@@ -68,6 +68,8 @@ We provide a variant of the TextInput that is a composed with CopyButton. In ord
 - To ensure `TextInput` meets accessibility standards, provide `name` and `id`
 - Keep in mind that to provide that accessibile inputs, you need to use labels.
 
+## Props
+
 import { Props } from '@contentful/f36-docs-utils';
 
 <Props of="TextInput" />
