--- packages/forma-36-react-components/src/components/Select/Option.tsx
@@ -1,20 +1,0 @@
-import React, { HTMLProps } from 'react';
-
-export interface OptionProps extends HTMLProps<HTMLOptionElement> {
-  value: string;
-  children: React.ReactNode;
-  testId?: string;
-}
-
-export const Option = ({
-  value,
-  children,
-  testId = 'cf-ui-select-option',
-  ...otherProps
-}: OptionProps) => {
-  return (
-    <option value={value} data-test-id={testId} {...otherProps}>
-      {children}
-    </option>
-  );
-};
