--- packages/forma-36-react-components/src/components/Select/index.ts
@@ -1,4 +1,0 @@
-export { Select } from './Select';
-export type { SelectProps } from './Select';
-export { Option } from './Option';
-export type { OptionProps } from './Option';
